+++ src/config.rs
@@ -15,6 +15,7 @@
 use crate::connectors::prelude::*;
 use simd_json::ValueType;
 use tremor_script::{
+    ast::deploy::ConnectorDefinition,
     ast::{self, Helper},
     FN_REGISTRY,
 };
@@ -152,7 +153,7 @@ pub(crate) struct Connector {
 
 impl Connector {
     /// Spawns a connector from a definition
-    pub(crate) fn from_defn(defn: &ast::ConnectorDefinition<'static>) -> crate::Result<Connector> {
+    pub(crate) fn from_defn(defn: &ast::ConnectorDefinition<'static>) -> crate::Result<Self> {
         let aggr_reg = tremor_script::registry::aggr();
         let reg = &*FN_REGISTRY.read()?;
 
@@ -161,55 +162,97 @@ impl Connector {
 
         let conf = params.generate_config(&mut helper)?;
 
-        Connector::from_config(defn.builtin_kind.clone().into(), &conf)
+        Self::from_config(defn.id.as_str(), defn.builtin_kind.clone().into(), &conf)
     }
     /// Creates a connector from it's definition (aka config + settings)
+    #[allow(clippy::too_many_lines)]
     pub(crate) fn from_config(
+        connector_id: &str,
         connector_type: ConnectorType,
         connector_config: &Value<'static>,
-    ) -> crate::Result<Connector> {
-        fn validate_type(v: &Value, k: &str, t: ValueType) -> Result<()> {
+    ) -> crate::Result<Self> {
+        fn validate_type(v: &Value, k: &str, t: ValueType, connector_id: &str) -> Result<()> {
             if v.get(k).is_some() && v.get(k).map(Value::value_type) != Some(t) {
-                return Err(format!(
-                    "Expect type {:?} for key {} but got {:?}",
-                    t,
-                    k,
-                    v.get(k).map_or(ValueType::Null, Value::value_type)
+                return Err(ErrorKind::InvalidConnectorDefinition(
+                    connector_id.to_string(),
+                    format!(
+                        "Expected type {t:?} for key {k} but got {:?}",
+                        v.get(k).map_or(ValueType::Null, Value::value_type)
+                    ),
                 )
                 .into());
             }
             Ok(())
         }
-        let config = connector_config.get("config").cloned();
+        let config = connector_config.get(ConnectorDefinition::CONFIG).cloned();
 
         // TODO: can we get hygenic errors here?
 
-        validate_type(connector_config, "codec_map", ValueType::Object)?;
-        validate_type(connector_config, "preprocessors", ValueType::Array)?;
-        validate_type(connector_config, "postprocessors", ValueType::Array)?;
-        validate_type(connector_config, "metrics_interval_s", ValueType::U64)
-            .or_else(|_| validate_type(connector_config, "metrics_interval_s", ValueType::I64))?;
+        validate_type(
+            connector_config,
+            ConnectorDefinition::CODEC,
+            ValueType::String,
+            connector_id,
+        )?;
+        validate_type(
+            connector_config,
+            ConnectorDefinition::CONFIG,
+            ValueType::Object,
+            connector_id,
+        )?;
+        validate_type(
+            connector_config,
+            ConnectorDefinition::RECONNECT,
+            ValueType::Object,
+            connector_id,
+        )?;
+        validate_type(
+            connector_config,
+            ConnectorDefinition::PREPROCESSORS,
+            ValueType::Array,
+            connector_id,
+        )?;
+        validate_type(
+            connector_config,
+            ConnectorDefinition::POSTPROCESSORS,
+            ValueType::Array,
+            connector_id,
+        )?;
+        validate_type(
+            connector_config,
+            ConnectorDefinition::METRICS_INTERVAL_S,
+            ValueType::U64,
+            connector_id,
+        )
+        .or_else(|_| {
+            validate_type(
+                connector_config,
+                ConnectorDefinition::METRICS_INTERVAL_S,
+                ValueType::I64,
+                connector_id,
+            )
+        })?;
 
-        Ok(Connector {
+        Ok(Self {
             connector_type,
             config,
             preprocessors: connector_config
-                .get_array("preprocessors")
+                .get_array(ConnectorDefinition::PREPROCESSORS)
                 .map(|o| o.iter().map(Preprocessor::try_from).collect::<Result<_>>())
                 .transpose()?,
             postprocessors: connector_config
-                .get_array("postprocessors")
+                .get_array(ConnectorDefinition::POSTPROCESSORS)
                 .map(|o| o.iter().map(Preprocessor::try_from).collect::<Result<_>>())
                 .transpose()?,
             reconnect: connector_config
-                .get("reconnect")
+                .get(ConnectorDefinition::RECONNECT)
                 .cloned()
                 .map(tremor_value::structurize)
                 .transpose()?
                 .unwrap_or_default(),
-            metrics_interval_s: connector_config.get_u64("metrics_interval_s"),
+            metrics_interval_s: connector_config.get_u64(ConnectorDefinition::METRICS_INTERVAL_S),
             codec: connector_config
-                .get("codec")
+                .get(ConnectorDefinition::CODEC)
                 .map(Codec::try_from)
                 .transpose()?,
         })
@@ -265,6 +308,7 @@ mod tests {
     #[test]
     fn test_config_builtin_preproc_with_config() -> Result<()> {
         let c = Connector::from_config(
+            "my_otel_client",
             ConnectorType::from("otel_client".to_string()),
             &literal!({
                 "preprocessors": [ {"name": "snot", "config": { "separator": "\n" }}],
@@ -280,6 +324,22 @@ mod tests {
                 assert_eq!("\n", config.get("separator").unwrap().to_string());
             }
         }
+        Ok(())
+    }
+
+    #[test]
+    fn test_connector_config_wrong_config() -> Result<()> {
+        let config = literal!({
+            "preprocessors": [],
+            "postprocessors": [],
+            "codec": "string",
+            "reconnect": {},
+            "metrics_interval_s": "wrong_type"
+        });
+        let id = "my_id";
+        let res = Connector::from_config(id, "fancy_schmancy".into(), &config);
+        assert!(res.is_err());
+        assert_eq!(String::from("Invalid Definition for connector \"my_id\": Expected type I64 for key metrics_interval_s but got String"), res.err().unwrap().to_string());
         Ok(())
     }
 }