+++ src/connectors/impls.rs
@@ -28,6 +28,8 @@ pub(crate) mod elastic;
 pub(crate) mod exit;
 /// file connector implementation
 pub(crate) mod file;
+/// Google Big Query
+pub(crate) mod gbq;
 /// HTTP
 pub(crate) mod http;
 /// Kafka consumer and producer