+++ src/errors.rs
@@ -184,6 +184,7 @@ error_chain! {
         WalInfailable(qwal::Error<std::convert::Infallible>);
         Uuid(uuid::Error);
         Serenity(serenity::Error);
+        InvalidMetadataValue(tonic::metadata::errors::InvalidMetadataValue);
     }
 
     errors {
@@ -273,6 +274,10 @@ error_chain! {
             description("Invalid Configuration")
                 display("Invalid Configuration for {}: {}", configured_thing, msg)
         }
+        InvalidConnectorDefinition(connector_id: String, msg: String) {
+            description("Invalid Connector Definition")
+                display("Invalid Definition for connector \"{}\": {}", connector_id, msg)
+        }
         InvalidConnect(target: String, port: Cow<'static, str>) {
             description("Invalid Connect attempt")
                 display("Invalid Connect to {} via port {}", target, port)
@@ -312,6 +317,18 @@ error_chain! {
         InvalidInputData(msg: &'static str) {
             description("Invalid Input data")
                 display("Invalid Input data: {}", msg)
+        }
+        GbqSinkFailed(msg: &'static str) {
+            description("GBQ Sink failed")
+                display("GBQ Sink failed: {}", msg)
+        }
+        BigQueryClientNotAvailable(msg: &'static str) {
+            description("BigQuery client not available")
+                display("BigQuery client not available: {}", msg)
+        }
+        BigQueryTypeMismatch(expected: &'static str, actual:value_trait::ValueType) {
+            description("Type in the message does not match BigQuery type")
+                display("Type in the message does not match BigQuery type. Expected: {}, actual: {:?}", expected, actual)
         }
     }
 }