+++ tremor-script/src/ast/deploy/raw.rs
@@ -166,6 +166,22 @@ impl_expr!(ConnectorDefinitionRaw);
 impl<'script> Upable<'script> for ConnectorDefinitionRaw<'script> {
     type Target = ConnectorDefinition<'script>;
     fn up<'registry>(self, helper: &mut Helper<'script, 'registry>) -> Result<Self::Target> {
+        // verify supported parameters
+        for (ident, _) in &self.params.with.exprs {
+            let key: &str = ident.id.as_ref();
+            if !ConnectorDefinition::AVAILABLE_PARAMS.contains(&key) {
+                let range = ident.mid.range;
+                return Err(ErrorKind::InvalidDefinitionalWithParam(
+                    range.expand_lines(2),
+                    range,
+                    format!("connector \"{}\"", self.id),
+                    ident.id.to_string(),
+                    &ConnectorDefinition::AVAILABLE_PARAMS,
+                )
+                .into());
+            }
+        }
+
         let query_defn = ConnectorDefinition {
             config: Value::const_null(),
             mid: self.mid.box_with_name(&self.id),
@@ -411,6 +427,27 @@ impl<'script> Upable<'script> for CreateStmtRaw<'script> {
                 }
             }
         };
+        let args = match defn {
+            CreateTargetDefinition::Connector(ref conn) => &conn.params.args.0,
+            CreateTargetDefinition::Pipeline(ref pipe) => &pipe.params.args.0,
+        };
+        for (ident, _) in &self.params.with.exprs {
+            if !args.iter().any(|(args_ident, _)| ident.id == args_ident.id) {
+                let range = ident.extent();
+                let available_args = args
+                    .iter()
+                    .map(|(ident, _)| ident.id.to_string())
+                    .collect::<Vec<String>>();
+                return Err(ErrorKind::WithParamNoArg(
+                    range.expand_lines(2),
+                    range,
+                    ident.id.to_string(),
+                    self.id.id.to_string(),
+                    available_args,
+                )
+                .into());
+            }
+        }
 
         let create_stmt = CreateStmt {
             mid: self.mid.box_with_name(&self.id.id),