+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 ### New features
 
 - Add support for `chunked` requests and responses for the HTTP connectors
-- 
+- Add the `gbq` connector for Google BigQuery
+
 ### Fixes
 
 - fix the use of `args` in the with part of a `create` insode of a `flow`.
+- fix silent swallowing of unknown fields in connector definition.
 
 ### Breaking Changes